pionex-bot/requirements.txt
httpx>=0.27.0
websockets>=12.0
python-dotenv>=1.0.0
numpy>=1.26.0
